존재하지 않는 게시물을 삭제할 경우, 오류 메세지

+++ src/main/java/com/kja/exam/jpaBoard/article/controller/ArticleController.java
@@ -51,6 +51,11 @@ public class ArticleController {
     @RequestMapping("doDelete")
     @ResponseBody
     public String doDelete(long id) {
+
+        if ( articleRepository.existsById(id) == false ) {
+            return "%d번 게시물은 이미 삭제되었거나 존재하지 않습니다.".formatted(id);
+        }
+
         articleRepository.deleteById(id);
         return "%d번 게시물이 삭제되었습니다.".formatted(id);
     }
